refactor: reorgaized and modularized routes and services
