+++ templates/auth/login.html
@@ -10,7 +10,7 @@
             <p class="text-gray-600">Enter your email to receive a login code</p>
         </div>
 
-        <form action="{{ url_for('request_otp') }}" method="POST">
+        <form action="{{ url_for('auth.request_otp') }}" method="POST">
             <div class="mb-6">
                 <label for="email" class="block text-sm font-medium text-gray-700 mb-2">
                     Email Address
