+++ templates/auth/verify.html
@@ -11,7 +11,7 @@
             <p class="text-emerald-600 font-semibold">{{ email }}</p>
         </div>
 
-        <form action="{{ url_for('verify_otp') }}" method="POST">
+        <form action="{{ url_for('auth.verify_otp') }}" method="POST">
             <input type="hidden" name="email" value="{{ email }}">
             
             <div class="mb-6">
@@ -39,7 +39,7 @@
         </form>
 
         <div class="mt-6 text-center">
-            <form action="{{ url_for('request_otp') }}" method="POST" style="display:inline;">
+            <form action="{{ url_for('auth.request_otp') }}" method="POST" style="display:inline;">
                 <input type="hidden" name="email" value="{{ email }}">
                 <button type="submit" class="text-emerald-600 hover:text-emerald-700 text-sm font-medium">
                     Didn't receive code? Resend
