+++ templates/base.html
@@ -38,7 +38,7 @@
                         </a>
                     </div>
                     {% else %}
-                    <a href="{{ url_for('login') }}" class="px-4 py-2 ml-4 rounded-lg bg-gradient-to-r from-emerald-600 to-teal-600 text-white hover:from-emerald-700 hover:to-teal-700 transition-all duration-200">
+                    <a href="{{ url_for('auth.login') }}" class="px-4 py-2 ml-4 rounded-lg bg-gradient-to-r from-emerald-600 to-teal-600 text-white hover:from-emerald-700 hover:to-teal-700 transition-all duration-200">
                         Login
                     </a>
                     {% endif %}
